LayoutTests/canvas/philip/tests/2d.text.measure.width.space.html

<!DOCTYPE html>
<title>Canvas test: 2d.text.measure.width.space</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<link rel="prev" href="2d.text.measure.width.empty.html" title="2d.text.measure.width.empty">
<link rel="next" href="index.html" title="index">
<style>
@font-face {
  font-family: CanvasTest;
  src: url("../fonts/CanvasTest.ttf");
}
</style><body class="show_output">
<p>
 <a href="2d.text.measure.width.empty.html" accesskey="p" title="[p] 2d.text.measure.width.empty">&lt;</a>
 <a href="index.html">[index]</a>
 <a href="index.html" accesskey="n" title="[n] index">&gt;</a>
<h1><a href="index.2d.html">2d</a>.<a href="index.2d.text.html">text</a>.<a href="index.2d.text.measure.html">measure</a>.<a href="index.2d.text.measure.width.html">width</a>.space</h1>
<p class="desc">Space characters are converted to U+0020 and collapsed (per CSS)</p>
<div class="refs">References:
<ul>
<li><a href="spec.html#testrefs.2d.text.measure.spaces">2d.text.measure.spaces</a>

</ul>
</div>

<span style="font-family: CanvasTest; position: absolute; visibility: hidden">A</span>
<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

deferTest();
setTimeout(wrapFunction(function () {
    ctx.font = '50px CanvasTest';
    _assertSame(ctx.measureText('A B').width, 150, "ctx.measureText('A B').width", "150");
    _assertSame(ctx.measureText('A  B').width, 150, "ctx.measureText('A  B').width", "150");
    _assertSame(ctx.measureText('A \x09\x0a\x0c\x0d  \x09\x0a\x0c\x0dB').width, 150, "ctx.measureText('A \\x09\\x0a\\x0c\\x0d  \\x09\\x0a\\x0c\\x0dB').width", "150");
    _assert(ctx.measureText('A \x0b B').width >= 200, "ctx.measureText('A \\x0b B').width >= 200");

    _assertSame(ctx.measureText(' AB').width, 100, "ctx.measureText(' AB').width", "100");
    _assertSame(ctx.measureText('AB ').width, 100, "ctx.measureText('AB ').width", "100");
}), 500);


});
</script>
