LayoutTests/canvas/philip/tests/2d.text.measure.width.space-expected.txt

< [index] >
2d.text.measure.width.space
Space characters are converted to U+0020 and collapsed (per CSS)
References: 2d.text.measure.spaces
Actual output:
Failed assertion ctx.measureText('A B').width === 150 (got 200[number], expected 150[number])
Failed assertion ctx.measureText('A \x09\x0a\x0c\x0d \x09\x0a\x0c\x0dB').width === 150 (got 450[number], expected 150[number])
Failed assertion ctx.measureText(' AB').width === 100 (got 150[number], expected 100[number])
Failed assertion ctx.measureText('AB ').width === 100 (got 150[number], expected 100[number])
